removed upcoder layers to save RAM

+++ rpt/testing.py
@@ -15,6 +15,6 @@ import torch
 
 from rpt.neox_model_torch import GPTNeoXForCausalLM
 
-hf_model = GPTNeoXForCausalLM.from_pretrained('Shahar603/neox-rpt-1-bf16', torch_dtype=torch.bfloat16)
+hf_model = GPTNeoXForCausalLM.from_pretrained('Shahar603/neox-rpt-1-bf16')
 encode_output = hf_model.encode(['hello'*100, 'world'], 1)
 print(encode_output)
